+++ core/embed/rust/src/ui/model_tt/component/homescreen/render.rs
@@ -463,7 +463,7 @@ impl BlurringContext {
         let data = get_data(buffer, self.line_num, mcu_height);
 
         for i in -BLUR_RADIUS..=BLUR_RADIUS {
-            let ic = i.clamp(0, HOMESCREEN_IMAGE_WIDTH as i16 - 1) as usize;
+            let ic = i.clamp(0, HOMESCREEN_IMAGE_WIDTH - 1) as usize;
             update_accs_add(data, ic, &mut acc_r, &mut acc_g, &mut acc_b);
         }
 
@@ -473,9 +473,9 @@ impl BlurringContext {
             self.lines[self.add_idx][BLUE_IDX][i as usize] = acc_b;
 
             // clamping handles left and right edges
-            let ic = (i - BLUR_RADIUS).clamp(0, HOMESCREEN_IMAGE_WIDTH as i16 - 1) as usize;
+            let ic = (i - BLUR_RADIUS).clamp(0, HOMESCREEN_IMAGE_WIDTH - 1) as usize;
             let ic2 = (i + BLUR_SIZE as i16 - BLUR_RADIUS)
-                .clamp(0, HOMESCREEN_IMAGE_WIDTH as i16 - 1) as usize;
+                .clamp(0, HOMESCREEN_IMAGE_WIDTH - 1) as usize;
             update_accs_add(data, ic2, &mut acc_r, &mut acc_g, &mut acc_b);
             update_accs_sub(data, ic, &mut acc_r, &mut acc_g, &mut acc_b);
         }
